Fix EXGeoHeader.version field width (was reading 2 packed u16s as one u32)
Disney Universe's GEOM/EDB version field reads as 131316 if treated as a
single u32, which is wildly outside the documented 182-263 range. The real
version (244, confirmed against the EngineX wiki's per-title version table)
is only the lower 16 bits; the upper 16 bits are a separate, still-unknown
field (observed value 2) that happens to be zero on every older title
eurochef already supports, which is why nobody had noticed the field is
really two u16s.

Splits EXGeoHeader.version into version_unk: u16 + version: u16, widens the
~7 downstream call sites that expected the old u32. With version correctly
read as 244, eurochef's edb commands get one field further into the real
header (erroring on refpointer_list instead of map_list) before hitting the
next unresolved issue: the list-region's actual byte offset still isn't
0x40 or 0x54 for this title, which needs more work (see project notes).

+++ eurochef/cli/src/edb/animations.rs
@@ -125,7 +125,7 @@ pub fn execute_command(
         edb.seek(std::io::SeekFrom::Start(a.common.address as u64))?;
 
         let skin = edb
-            .read_type_args::<EXGeoBaseAnimSkin>(edb.endian, (header.version,))
+            .read_type_args::<EXGeoBaseAnimSkin>(edb.endian, (header.version as u32,))
             .context("Failed to read animation")?;
 
         let entity_indices: Vec<u32> = skin
@@ -145,7 +145,7 @@ pub fn execute_command(
 
             edb.seek(std::io::SeekFrom::Start(e.common.address as u64))?;
 
-            let ent = edb.read_type_args::<EXGeoEntity>(edb.endian, (header.version, platform));
+            let ent = edb.read_type_args::<EXGeoEntity>(edb.endian, (header.version as u32, platform));
 
             if let Err(err) = ent {
                 error!("Failed to read entity: {err}");
